+++ src/Exceptions/AudienceMismatchException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jmluang\SsoConsumer\Exceptions;
+
+class AudienceMismatchException extends SsoConsumerException
+{
+    public const ERROR_CODE = 'audience_mismatch';
+}