+++ src/Exceptions/ReplayedTicketException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jmluang\SsoConsumer\Exceptions;
+
+class ReplayedTicketException extends SsoConsumerException
+{
+    public const ERROR_CODE = 'ticket_replayed';
+}