+++ src/Exceptions/ResolverFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jmluang\SsoConsumer\Exceptions;
+
+class ResolverFailedException extends SsoConsumerException
+{
+    public const ERROR_CODE = 'resolver_failed';
+}