+++ src/Exceptions/TenantMismatchException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jmluang\SsoConsumer\Exceptions;
+
+class TenantMismatchException extends SsoConsumerException
+{
+    public const ERROR_CODE = 'tenant_mismatch';
+}